requirements.txt
opencv-python==4.8.1.78
pyaudio==0.2.13
Pillow==10.1.0
numpy==1.24.3